+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Version 0.7.0: major bugfix and API change release
 
 - Upgrade to latest version of `epfl_si.actions`, thereby gaining `delegate_to:` support in all tasks
+- Rewrite documentation using the `ansible-doc` standard
 - Drop `epfl_si.rancher.rke2_access` role, as it was unused and at odds with 0.6.0's strategy of not managing kubeconfig files anymore
 - `epfl_si.rancher.rke2-node` role
 - `epfl_si.rancher.rancher_project` lookup plugin